requirements.txt
annotated-types==0.7.0
anyio==4.6.0
bcrypt==4.2.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
cryptography==43.0.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
fastapi==0.115.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
MarkupSafe==3.0.1
numpy==2.0.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pluggy==1.5.0
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pytest==8.3.3
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-multipart==0.0.12
pytz==2024.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.39.2
tomli==2.0.2
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.31.1
